+++ modules/cmsadmin/apis/NavItemController.php
@@ -349,6 +349,7 @@ class NavItemController extends \admin\base\RestController
             'cfgs' => $blockObject->getCfgs(),
             'extras' => $blockObject->extraVars(),
             'values' => $blockItem['json_config_values'],
+            'field_help' => $blockObject->getFieldHelp(),
             'cfgvalues' => $blockItem['json_config_cfg_values'], // add: t1_json_config_cfg_values
             '__placeholders' => $placeholders,
         ];